Restructure transformer
